refactor(sites): 抽取站点指标聚合到 utils/siteMetrics.js
